+++ typescript/packages/mechanisms/evm/src/upto/facilitator/index.ts
@@ -1,3 +1,1 @@
 export { UptoEvmScheme } from "./scheme";
-export { registerUptoEvmScheme } from "./register";
-export type { UptoEvmFacilitatorConfig } from "./register";
