+++ typescript/packages/mechanisms/evm/src/upto/server/index.ts
@@ -1,3 +1,1 @@
 export { UptoEvmScheme } from "./scheme";
-export { registerUptoEvmScheme } from "./register";
-export type { UptoEvmResourceServerConfig } from "./register";
